+++ docs/conf.py
@@ -55,7 +55,7 @@ import os
 # directory, add these directories to sys.path here. If the directory is
 # relative to the documentation root, use os.path.abspath to make it
 # absolute, like shown here.
-#sys.path.insert(0, os.path.abspath('.'))
+sys.path.insert(0, os.path.abspath('.'))
 
 # Get the project root dir, which is the parent dir of this
 cwd = os.getcwd()
@@ -75,7 +75,13 @@ import hydrofunctions
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = ['sphinx.ext.autodoc', 'sphinx.ext.viewcode', 'sphinx.ext.autosectionlabel', 'sphinxcontrib.napoleon']
+extensions = [
+        'sphinx.ext.autodoc',
+        'sphinx.ext.viewcode',
+        'sphinx.ext.autosectionlabel',
+        #'sphinxcontrib.napoleon',
+        'sphinx.ext.napoleon',
+        ]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']