+++ tests/test_station.py
@@ -14,13 +14,16 @@ import numpy as np
 
 from hydrofunctions import station, typing
 from .fixtures import (
-        fakeResponse
+        fakeResponse,
+        recent_only,
         )
 
 
 class TestingNWIS(station.NWIS):
     """
-    This subclass of NWIS is for testing the NWIS methods.
+    This subclass of NWIS is for testing all of the NWIS methods except
+    __init__, which we will replace. All of the other methods get inherited
+    verbatim, so we can test them using TestingNWIS instead of NWIS.
     """
     def __init__(self, site=None, service=None, start_date=None, end_date=None, dataframe=None, meta=None, start=None, end=None):
         self.site = site
@@ -88,7 +91,7 @@ class TestStation(unittest.TestCase):
         self.assertIsInstance(actual, Foo)
 
 
-class TestNWIS(unittest.TestCase):
+class TestNWISinit(unittest.TestCase):
 
     @mock.patch("hydrofunctions.hydrofunctions.get_nwis")
     @mock.patch("hydrofunctions.hydrofunctions.get_nwis_property")
@@ -184,6 +187,35 @@ class TestNWIS(unittest.TestCase):
         self.assertEqual(expected_meta, actual.meta, 'The metadata were not retrieved by NWIS.read().')
 
 
+    @mock.patch("hydrofunctions.hydrofunctions.get_nwis")
+    def test_NWIS_init_request_most_recent_only(self, mock_get_nwis):
+        expected_json = recent_only
+        expected_url = 'https://waterservices.usgs.gov/nwis/dv/?format=json%2C1.1&sites=01541200'
+        mock_get_nwis.return_value = fakeResponse(json=expected_json, url=expected_url)
+        actual = station.NWIS('01541200')
+        self.assertEqual(actual.df().shape, (2,4),'The dataframe should only have two rows and four columns.')
+
+
+class TestNWISmethods(unittest.TestCase):
+    """
+    Tests for NWIS methods
+    The following section is for testing all of the NWIS methods
+    EXCEPT the NWIS.__init__() method.
+
+    Creating an NWIS instance will always run the __init__ method, which we
+    usually don't want to do. It calls a bunch of functions that we test
+    elsewhere and it causes a bunch of side effects that we don't want. Yes,
+    you can mock all of the functions that __init__ calls, but even then there
+    can be unwanted side effects not to mention it can be tedious to mock so
+    many different things.
+
+    To test any method other than __init__, we will use the following strategy:
+        - create a sub-class of NWIS called TestingNWIS.
+        - TestingNWIS has a different __init__ method that allows you to pass
+            in a dataframe and any other initial parameter
+        - all other methods gets inherited from NWIS, so we can test them.
+"""
+
     def test_NWIS_df_returns_all_columns(self):
         expected_cols = ['USGS:01541200:00060:00000_qualifiers',
                          'USGS:01541200:00060:00000',
@@ -218,6 +250,29 @@ class TestNWIS(unittest.TestCase):
         actual_df = test_nwis.df('all')
         actual_cols = actual_df.columns.tolist()
         self.assertListEqual(actual_cols, expected_cols, "NWIS.df() should return all of the columns.")
+
+    def test_NWIS_df_data_returns_data_columns(self):
+        cols = ['USGS:01541200:00060:00000_qualifiers',
+                         'USGS:01541200:00060:00000',
+                         'USGS:01541200:00065:00000_qualifiers',
+                         'USGS:01541200:00065:00000',
+                         'USGS:01541303:00060:00000_qualifiers',
+                         'USGS:01541303:00060:00000',
+                         'USGS:01541303:00065:00000_qualifiers',
+                         'USGS:01541303:00065:00000']
+        expected_cols = ['USGS:01541200:00060:00000',
+                         'USGS:01541200:00065:00000',
+                         'USGS:01541303:00060:00000',
+                         'USGS:01541303:00065:00000']
+
+        data = [['test', 5, 'test', 5, 'test', 5, 'test', 5],
+                ['test', 5, 'test', 5, 'test', 5, 'test', 5]]
+        test_df = pd.DataFrame(data=data, columns=cols)
+        test_nwis = TestingNWIS(dataframe=test_df)
+        actual_df = test_nwis.df('data')
+        actual_cols = actual_df.columns.tolist()
+        self.assertListEqual(actual_cols, expected_cols, "NWIS.df('data') should return all of the data columns.")
+
 
     def test_NWIS_df_discharge_returns_discharge_data_columns(self):
         cols = ['USGS:01541200:00060:00000_qualifiers',